# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Ethnos.app Academic Bibliography API
  version: 2.0.0
  description: "

    REST API for academic bibliography: search, works, persons,
    organizations, venues, courses, citations, collaborations and health.


    Highlights

    - Public endpoints; selected admin/metrics routes require
    'x-access-key'.

    - Standard response envelope: { status, data, pagination?, meta? }.

    - Pagination supports page/limit and offset/limit.

    - Identifiers: DOI, ORCID, ROR, ISSN.


    Notes

    - Rate limiting is configurable via environment and enabled by default.

    - Performance and dataset size vary by deployment.

    \        "
  contact:
    name: Bruno Cesar Cunha Cruz, PhD Student
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  x-developer-orcid: 0000-0001-8652-2333
  x-institution: PPGAS/MN/UFRJ
servers:
  - url: https://api.ethnos.app
    description: Production API Server - Enterprise-ready academic research infrastructure
  - url: http://localhost:1210
    description: Development Server - Local testing and development environment
...
components:
  securitySchemes:
    XAccessKey:
      type: apiKey
      in: header
      name: x-access-key
      description: Internal access key required for protected endpoints (security,
        dashboard, health except /live)
  schemas:
...
    Error:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: Resource not found
        code:
          type: string
          example: NOT_FOUND
        timestamp:
          type: string
          format: date-time
...
    Person:
      type: object
      properties:
        id:
          type: integer
          example: 5952
          description: Unique identifier for the person
        preferred_name:
          type: string
          example: Dr. Maria Silva Santos
          description: Preferred display name of the person
        given_names:
          type: string
          example: Maria Silva
          description: Given names
        family_name:
          type: string
          example: Santos
          description: Family name or surname
        orcid:
          type: string
          example: 0000-0002-1825-0097
          description: ORCID identifier
        lattes_id:
          type: string
          example: "1234567890123456"
          description: Lattes CV platform ID (Brazil)
        scopus_id:
          type: string
          example: "57194582100"
          description: Scopus Author ID
        wos_id:
          type: string
          example: A-1234-2023
          description: Web of Science ResearcherID
        primary_affiliation:
          $ref: "#/components/schemas/Organization"
          description: Primary institutional affiliation
        works_count:
          type: integer
          example: 45
          description: Total number of works authored
        h_index:
          type: integer
          example: 12
          description: H-index bibliometric indicator
        citation_count:
          type: integer
          example: 678
          description: Total citations received
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  responses:
...
    BadRequest:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: error
            message: Invalid venue ID
            code: VALIDATION_ERROR
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: error
            message: Resource not found
            code: NOT_FOUND
...
paths:
...
  "/persons/{id}/signatures":
    get:
      summary: Get signatures linked to a person
      description: Retrieve all name signatures associated with a specific person
      tags:
        - Persons
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
          description: Person ID
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Results per page
      responses:
        "200":
          description: Signatures retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: Signature ID
                        signature:
                          type: string
                          description: Name signature
                        created_at:
                          type: string
                          format: date-time
                        persons_count:
                          type: integer
                          description: Total persons with this signature
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
                      hasNext:
                        type: boolean
                      hasPrev:
                        type: boolean
        "404":
          description: Person not found
        "500":
          description: Internal server error
...
  /signatures:
    get:
      summary: Get list of name signatures
      description: Retrieve a paginated list of author name signatures with filtering
        and sorting options
      tags:
        - Signatures
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page (max 100)
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of items to skip
        - in: query
          name: search
          schema:
            type: string
            minLength: 1
            maxLength: 100
          description: Search term to filter signatures
          example: Smith
        - in: query
          name: sortBy
          schema:
            type: string
            enum:
              - signature
              - created_at
              - id
          description: Field to sort by
          example: signature
        - in: query
          name: sortOrder
          schema:
            type: string
            enum:
              - ASC
              - DESC
          description: Sort order
          example: ASC
        - in: query
          name: light
          schema:
            type: boolean
            default: false
          description: If true, omits expensive counts to speed up listing
      responses:
        "200":
          description: List of signatures retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Signature"
                  pagination:
                    type: object
                    properties:
                      limit:
                        type: integer
                        example: 20
                      offset:
                        type: integer
                        example: 0
                      total:
                        type: integer
                        example: 378134
                      hasNext:
                        type: boolean
                        example: true
                      hasPrev:
                        type: boolean
                        example: false
        "400":
          $ref: "#/components/responses/BadRequest"
  /signatures/statistics:
    get:
      summary: Get signature statistics
      description: Retrieve comprehensive statistics about signatures in the system
      tags:
        - Signatures
      responses:
        "200":
          description: Signature statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      total_signatures:
                        type: integer
                        example: 378134
                      average_signature_length:
                        type: number
                        format: float
                        example: 10.17
                      total_characters:
                        type: integer
                        example: 3847823
                      shortest_signature:
                        type: integer
                        example: 1
                      longest_signature:
                        type: integer
                        example: 255
        "500":
          description: Internal server error
  /signatures/search:
    get:
      summary: Search signatures
      description: Search for signatures by name with exact and fuzzy matching options
      tags:
        - Signatures
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 100
          description: Search query for signature name
          example: J. Smith
        - in: query
          name: exact
          schema:
            type: boolean
            default: false
          description: Whether to perform exact matching
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of items to skip
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Signature"
                  pagination:
                    type: object
                  query:
                    type: string
                    example: J. Smith
                  exact_match:
                    type: boolean
                    example: false
        "400":
          $ref: "#/components/responses/BadRequest"
  "/signatures/{id}":
    get:
      summary: Get signature by ID
      description: Retrieve detailed information about a specific signature
      tags:
        - Signatures
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
          description: Signature ID
          example: 123
      responses:
        "200":
          description: Signature details
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: "#/components/schemas/Signature"
        "404":
          $ref: "#/components/responses/NotFound"
  "/signatures/{id}/persons":
    get:
      summary: Get persons associated with a signature
      description: Retrieve all persons associated with a specific signature
      tags:
        - Signatures
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
          description: Signature ID
          example: 123
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of items to skip
      responses:
        "200":
          description: Associated persons
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Person"
                  pagination:
                    type: object
        "404":
          $ref: "#/components/responses/NotFound"
  "/signatures/{id}/works":
    get:
      summary: Get works associated with a signature
      description: Retrieve all academic works (publications) associated with a
        specific signature
      tags:
        - Signatures
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
          description: Signature ID
          example: 92152
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Results per page
      responses:
        "200":
          description: Works retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: Work ID
                        title:
                          type: string
                          description: Work title
                        subtitle:
                          type: string
                          description: Work subtitle
                        type:
                          type: string
                          description: Type of work
                        language:
                          type: string
                          description: Publication language
                        doi:
                          type: string
                          description: DOI identifier
                        authorship:
                          type: object
                          properties:
                            role:
                              type: string
                              enum:
                                - AUTHOR
                                - EDITOR
                              description: Person's role in this work
                            position:
                              type: integer
                              description: Position in author list
                            is_corresponding:
                              type: boolean
                              description: Whether this person is corresponding author
                            person_id:
                              type: integer
                              description: Person ID
                            person_name:
                              type: string
                              description: Person name
                        publication:
                          type: object
                          properties:
                            year:
                              type: integer
                              description: Publication year
                            journal:
                              type: string
                              description: Journal name
                            volume:
                              type: string
                              description: Volume number
                            issue:
                              type: string
                              description: Issue number
                            pages:
                              type: string
                              description: Page range
                        authors:
                          type: object
                          properties:
                            total_count:
                              type: integer
                              description: Total number of authors
                            author_string:
                              type: string
                              description: Full author list string
                        created_at:
                          type: string
                          format: date-time
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
                      hasNext:
                        type: boolean
                      hasPrev:
                        type: boolean
        "404":
          description: Signature not found
        "500":
          description: Internal server error
...